+++ lib/solargraph/language_server/message/initialize.rb
@@ -36,6 +36,7 @@ module Solargraph
           result[:capabilities].merge! static_references unless dynamic_registration_for?('textDocument', 'references')
           result[:capabilities].merge! static_workspace_symbols unless dynamic_registration_for?('workspace', 'symbol')
           result[:capabilities].merge! static_folding_range unless dynamic_registration_for?('textDocument', 'foldingRange')
+          result[:capabilities].merge! static_highlights unless dynamic_registration_for?('textDocument', 'documentHighlight')
           # @todo Temporarily disabled
           # result[:capabilities].merge! static_code_action unless dynamic_registration_for?('textDocument', 'codeAction')
           set_result result
@@ -135,6 +136,12 @@ module Solargraph
           return {} unless host.options['folding']
           {
             foldingRangeProvider: true
+          }
+        end
+
+        def static_highlights
+          {
+            documentHighlightProvider: true
           }
         end